Delete on blur trouble prevented click
Edit a todo and blank the title. Then blur by clicking the complete on
another todo for example. This should delete the todo and complete the
other todo. But the first edit was cancelled by the second request. This
works for a normal edit but for a delete the counts must be updated. So
changed to call the delete Action client side if the title's blank which
won't get cancelled. But this caused the second response to be cancelled
because the Url was changed by the delete response. Changed the delete
action to leave the Url unchanged by not blanking the id.